+++ src/MacroTools/Cheats/CheatBuild.cs
@@ -36,7 +36,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
 
       var enteredString = GetEventPlayerChatString();
       var p = GetTriggerPlayer();
