+++ src/MacroTools/Cheats/CheatGold.cs
@@ -8,7 +8,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
       string enteredString = GetEventPlayerChatString();
       string parameter = null;
       player p = GetTriggerPlayer();
