+++ src/MacroTools/Cheats/CheatHasResearch.cs
@@ -15,7 +15,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
 
       var enteredString = GetEventPlayerChatString();
       var parameter = SubString(enteredString, StringLength(Command), StringLength(enteredString));
