+++ src/MacroTools/Cheats/CheatKick.cs
@@ -13,7 +13,7 @@ namespace MacroTools.Cheats
     {
       try
       {
-        if (!TestSafety.CheatCondition()) return;
+        if (!TestMode.CheatCondition()) return;
 
         string enteredString = GetEventPlayerChatString();
         player p = GetTriggerPlayer();
