+++ src/MacroTools/Cheats/CheatLumber.cs
@@ -12,7 +12,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
       string enteredString = GetEventPlayerChatString();
       string parameter = null;
       player p = GetTriggerPlayer();
