+++ src/MacroTools/Cheats/CheatMP.cs
@@ -16,7 +16,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
       string enteredString = GetEventPlayerChatString();
       player p = GetTriggerPlayer();
       _parameter = SubString(enteredString, StringLength(Command), StringLength(enteredString));
