+++ src/MacroTools/Cheats/CheatOwner.cs
@@ -15,7 +15,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
 
       string enteredString = GetEventPlayerChatString();
       player p = GetTriggerPlayer();
