+++ src/MacroTools/Cheats/CheatSpawn.cs
@@ -18,7 +18,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
       var enteredString = GetEventPlayerChatString();
       var triggerPlayer = GetTriggerPlayer();
       var typeIdParameter = SubString(enteredString, StringLength(Command), StringLength(Command) + 4);
