+++ src/MacroTools/Cheats/CheatTele.cs
@@ -20,7 +20,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
 
       string enteredString = GetEventPlayerChatString();
       player p = GetTriggerPlayer();
