+++ src/MacroTools/Cheats/CheatVision.cs
@@ -10,7 +10,7 @@ namespace MacroTools.Cheats
 
     private static void Actions()
     {
-      if (!TestSafety.CheatCondition()) return;
+      if (!TestMode.CheatCondition()) return;
       string enteredString = GetEventPlayerChatString();
       player p = GetTriggerPlayer();
       string parameter = SubString(enteredString, StringLength(Command), StringLength(enteredString));
