+++ src/WarcraftLegacies.Source/Setup/CheatSetup.cs
@@ -8,7 +8,7 @@ namespace WarcraftLegacies.Source.Setup
   {
     public static void Setup()
     {
-      TestSafety.Setup();
+      TestMode.Setup();
       CheatBuild.Setup();
       CheatControl.Setup();
       CheatFaction.Setup();
